# docs/implementation/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pongdang API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/water-index/assessments": {
      "get": {
        "tags": [
          "water-index"
        ],
        "summary": "Assessments",
        "operationId": "assessments_api_data_water_index_assessments_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "title": "Activity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Until"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "const": "general",
              "type": "string",
              "title": "Profile Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "observation",
                "forecast"
              ],
              "type": "string",
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/water-forecast/forecasts": {
      "get": {
        "tags": [
          "water-forecast"
        ],
        "summary": "Forecasts",
        "operationId": "forecasts_api_data_water_forecast_forecasts_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "title": "Activity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Until"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "general",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "observation",
                    "forecast"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/water-twin": {
      "get": {
        "tags": [
          "water-twin"
        ],
        "summary": "Twin",
        "operationId": "twin_api_data_water_twin_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "default": "swim",
              "title": "Activity"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "west",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "West"
            }
          },
          {
            "name": "east",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "East"
            }
          },
          {
            "name": "south",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90,
                  "minimum": -90
                },
                {
                  "type": "null"
                }
              ],
              "title": "South"
            }
          },
          {
            "name": "north",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90,
                  "minimum": -90
                },
                {
                  "type": "null"
                }
              ],
              "title": "North"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "observation",
                "forecast"
              ],
              "type": "string",
              "default": "observation",
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpatialEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "water-assessment.v1-draft",
            "title": "Contract Version",
            "default": "water-assessment.v1-draft"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "query": {
            "$ref": "#/components/schemas/ReadQuery"
          },
          "coverage": {
            "$ref": "#/components/schemas/Coverage"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/AssessmentDTO"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of",
          "queried_at",
          "query",
          "coverage",
          "rows",
          "total",
          "page",
          "page_size"
        ],
        "title": "Envelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpatialEnvelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "water-spatial.v1",
            "title": "Contract Version",
            "default": "water-spatial.v1"
          },
          "query": {
            "$ref": "#/components/schemas/SpatialQuery"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "no_data"
            ],
            "title": "Status"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/SpatialPlace"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Codes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "as_of",
          "at",
          "status",
          "rows",
          "has_more",
          "reason_codes"
        ],
        "title": "SpatialEnvelope"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}